refactor: BookVO 변수 수정, BookService 수정( getInfo -> detailBookInfo), main.jsp 이미지 띄우는 분기 처리(기존에 있던 사진과 직접 넣은사진), details.jsp 이미지 띄우는 분기 처리 등

+++ meta-bookdstore/src/main/java/com/meta/book/service/BookService.java
@@ -19,11 +19,11 @@ public class BookService {
 		return mapper.list();
 	}
 
-	public void delete(int book_no){
+	public void delete(Long book_no){
 		mapper.delete(book_no);
 	}
 
-	public List<BookVO> detailBookInfo(int book_no){
+	public List<BookVO> detailBookInfo(Long book_no){
 
 		return mapper.detailBookInfo(book_no);
 	}
